requirements.txt
django==3.1.5
djangorestframework==3.12.2
mysqlclient==1.4.2